roman_senate_game/config.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-

"""
Roman Senate AI Game
Configuration Module

This module handles configuration settings for the application,
including API keys and environment variables.
"""

import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

# OpenAI API configuration
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")